Refactor the Postgres service.
Refactored the Postgres service to better resemble an existing setup.
This means that we now have a default "postgres" user that creates an
additional Debezium user.

Additionally, the debezium test tables have been updated so that we
have one with a primary key and one without so that we can test both
scenarios.  Table 1 has "REPLICA IDENTITY FULL" because for whatever
reason, the prior state of an UPDATE is not included without it even
though the docs say the primary key should be enough.  Table 2 also has
"REPLICA IDENTITY FULL" because it needs some way to track the data
without having a primary key.